KTOR-2472 Extract events to shared module
