+++ ktor-server/ktor-server-core/jvm/src/io/ktor/features/CallLogging.kt
@@ -5,6 +5,7 @@
 package io.ktor.features
 
 import io.ktor.application.*
+import io.ktor.events.*
 import io.ktor.http.*
 import io.ktor.request.*
 import io.ktor.util.*
@@ -19,7 +20,7 @@ import kotlin.coroutines.*
  */
 public class CallLogging private constructor(
     private val log: Logger,
-    private val monitor: ApplicationEvents,
+    private val monitor: Events,
     private val level: Level,
     private val filters: List<(ApplicationCall) -> Boolean>,
     private val mdcEntries: List<MDCEntry>,
