+++ settings.gradle
@@ -103,3 +103,4 @@ include ':ktor-shared:ktor-shared-serialization'
 include ':ktor-shared:ktor-shared-serialization-kotlinx'
 include ':ktor-shared:ktor-shared-serialization-gson'
 include ':ktor-shared:ktor-shared-serialization-jackson'
+include ':ktor-shared:ktor-shared-events'
